+++ NEWS.md
@@ -1,6 +1,7 @@
 # PEF 3.2.3
 - Port `max_flt_nan` and `max_par_nan` parameters to R function `compute_analogs`.
 - Update R and C++ document.
+- Add R configuration `preserve_search_stations` to `independentSearch` as well. Now, both `independentSearch` and `extendedSearch` support this configuration.
 
 # PEF 3.2.2
 - Improve backward compatibility of R function `compute_analogs`. This function now returns the analog member index in the 4D observations, rather than in the 3D observations.